+++ utilities/sample-data/global-navigation-bar/index.jsx
@@ -74,7 +74,6 @@ const propSets = {
 	noNav: {
 		props: {},
 		primaryRegionProps: {
-			dividerPosition: null,
 			appLauncher: {
 				assistiveText: 'Open App Launcher',
 				id: 'app-launcher-trigger',